reminders without cron, exec() or wget :D
For quite some time, I desired to implement this tiny feature
now, the reminders run in the background, but without the help
of cron, wget, or the exec()'uted PHP binary.

see the source for the trick.

ps: this is the first , and last feature I 'll implement for 0.9.9
next commit, is the the installer's turn (well. or maybe fixes for the bugs generated
by this change :D)



git-svn-id: https://flyspray.svn.sourceforge.net/svnroot/flyspray/branches/0.9.9@800 2633ddf7-78cb-0310-bf30-a8c8deaca200