more jabber improvements ( florian check this please;))
more improvements on the jabber stuff to make it work and
less misteriuos to the user.


schedule.php --> actually daemonize, no. I mean, ***really**
daemonize it, and yes, it does work :D please check the developer
warning notes before debugging.

ok. class.fsjabber* --> class.fsjabberinfo.php gathers information
to be displayed to the user, about the capabilites he can use
the GetInfo() method abstracts the gory details and returns a
array where:

$results['enabled'] array  shows the features the user can use.
$results['disabled'] array shows the features the user cannot use

this should be disaplayed in a nice box in the left of the admin jabber
server settings and documented into the wiki for people finding more
clues.

so

require_once BASEDIR . '/includes/class.fsjabberinfo.php';

$info = new fsJabberInfo;

foreach($info->GetInfo() as $key=>$value) {

/*do something nice and eye candy to display that information
pointing to the wiki url with the description, otherwise, this will
stays as a mistery
*/
}

if you like this in other way. feel free to change it.

and don't blame me if you don't like the requirements. that's simple how
the things works  :-(

1. if no openssl in PHP you cannot use any jabber ssl or TLS or any
other thing requiring advanced crypto .

2. if no linux and PHP5 you cannot use SRV support, sorry. this is a
limitation of the language provided functions, see the dns_get_record
manual page in php.net.

3. you cannot use digest auth if you don't have mhash loaded, sorry.
this is a limitation of the jabber class that **may** be possible to
change, I haven't checked the alternatives yet.



git-svn-id: https://flyspray.svn.sourceforge.net/svnroot/flyspray/branches/0.9.9@810 2633ddf7-78cb-0310-bf30-a8c8deaca200