Minor corrections
add BASEDIR usage to feeds.php

add a workaround to circunvent
http://www.hardened-php.net/hphp/zend_hash_del_key_or_index_vulnerability.html
on systems where register_globals is enabled,  'cause is really
dangerous with such configuration.

Note that flyspray do an important effort to eliminate register_globals
annoyances, but you are on your own if you insist in configuring your
server in such insecure way.

Users with register_globals off are still vulnerable to this problem,
but in a less dangerous way, there is no 100% effective workaround in
this case, you must upgrade your PHP version.

1.rework some tiny parts of the class project, no functionaly change.

2. added a method to detect the mime type of the uploaded files is a
more reliable way, only the method is commited now, the funcionality
will be commited later,

3.improve filters class.



git-svn-id: https://flyspray.svn.sourceforge.net/svnroot/flyspray/branches/0.9.9@834 2633ddf7-78cb-0310-bf30-a8c8deaca200