Double e-mail input
You know what? The FIRST user to register on my flyspray instance wrote
his email address wrongly, and I couldn't resend him a code from the
administration pannel. So I had to delete him, it was bugged... With
double e-mail input we usually lower the risks of typos in e-mails