Fix estimated_effort string mess
Peterdd is right, the use of "estimated_effort" is only n the tasklist.
Now, this fix has pros and cons...

I thought that we should keep "_" of the language file as it was made
until now, so we can't put "estimated_effort" in there. But the column
form DB comes with an underscore... And in the DB we separate words with
underscores, so I can't change it there either... So the solution is to
str_replace() it. Happily waiting on your comment if you have another
idea...