added the 3 new translations for massop toggle
I'm a bit unsure about capitalize words directly in lang/en.php in english. In principle the translations can be kept lowercase. Then, where the string appears in the HTML, the CSS rule text-transform:capitalize; can be used on the appropriate CSS class.

With that given, it is in principle possible to reuse one translation string at different parts. At one position (something like a heading) we can use the transform:capitalize; 
- and at another position it is better to keep it lowercase.

In german nouns are always capitalized, while verbs and adjectives are lowercase (at normal cases).
Some words can mean noun or verb ( Regeln (noun) vs regeln  (wir regeln = we rule) which in english would be rules (the rules) vs rules (he rules). So for germans this is more important for fluently reading and understanding grammar.

On headings in german we only capitalize the first letter.
On headings in english every word seems to be capitalized (thats what I learned 25 years ago in school, but that were only for writing on paper) But in the web .. "Commit changes" on github.com seems also only to capitalize the first word of heading...