don't build answer page content before forms are handled
When user a form submits, the result can change also viewable data on the answer page. So the form must be handled first with safing data to database and then collect the data from the database to generate the answer page.

The alternative of redirecting to a second page isn't nice and can lead to harder to detect issues.