Fix for Composer installer part 1
Well part 2 is on you @peterdd I cannot test the .pl 's

What that fix is about:

- If we detect that vendors need installation, composerit.txt is copied
over to root directory.
- When running composerit.php, it will download composer using the php
command, as we  are sure php is running  on the machine. Not the case
with curl.
- composer.phar will now be downloaded to the right place, the root
directory, where the second cmd will run it, and it will find the json
file and thus do download the vendors where they belong.
- Previously they weren't downloaded as composer did not find any json
file in setup directory