fix FS#1841
The reason was a javascript loaded on this page, that autolinks to the task of the clicked row. well on user  account related events there is not task id..

I think this script can be replaced by html only.

But for this bugfix we simply change the id of the eventlog html table.