severity or priority is set to 0 by default
In the case there exists severity or priority entries in tasks with db default values of 0 we get a warning.

This is a quick fix too.

Maybe we have to check later (FS1.1) if a existing severtiy or priority level is deleted that have task assigned to that level. ->classic foreign key problem. Should be "ON DELETE RESTRICT". "ON DELETE SET DEFAULT" would move the tasks to default priority 0 or default severity 0.