use of project id for cache id construction
If you call for examples listTaskTypes(true) for different "projects" in one request (like $project1 and $project0 for the global values), 
it returned always only the result of the first query.

I need this to be fixed working on the field lists of project management pages for FS1.0. Want to show the global list values so project managers don't accidently create duplicate/similiar list field values.