task list and roadmap menu links only if allowed
NEW FEATURE: For roadmap guest view permission we probably need a new project setting. For me it would be ok to user others_view option, but there are probably use cases where guests can view the tasks but not the roadmap.

For a public project bug tracking site like bugs.flyspray.org the view of roadmap should be visible also for guests.