project css theming stub
adds the custom project css at the right position in header.tpl.

This is one step of compensation that there is only one theme structure (CleanFS) is maintained for FS1.0 and no other themes like the old classic blueprint theme.

The idea is you can customize your Flyspray and each project by adding a custom_xxx.css to the CleanFS/ directory and it is then available as custom css style. 

TODO: select dropdown menu in admin.prefs and pm prefs for setting a custom css for a project or flyspray global views.