consistent linking to users page
I think it is better just to always link to the user's page instead direct to the 'edit user' page also for admins.
With more personalisation coming (personal todo list, charts, ..) in future to flyspray
the user page itself is of increasing value. Instead show the 'edit user'-button on user's page.