Bugfix with quickediting estimate_efforts
This little bugfix repairs the quickediting for estimate_efforts. It's kind a little bit dirty, I think, because the quickedit-functions doesn't have a "do"-function. The edit of the header-file is needed for the quickedit.php on line 46-48. This lines need a correct project-Object. Previously the default-project was choosen there...