adding floeles dbtable comments to flyspray-install.xml
The <descr> tag can be used for commenting table and fields, see xmlschema03.dtd of AdoDB.

The mysldump I got from floele a year ago included sql table comments, but not existend in Flyspray github source.
So then just lets add them to flyspray-install.xml too.

We can enhance and completion can be done.

Note: <descr> tags in xmlschemas are currently ignored by adodb 5.20.4. (because not yet implemented, but you can vote for 
https://github.com/ADOdb/ADOdb/issues/256 it. ) :-)
Purpose is having DBschema documention in the Flyspray source too. And in future also in the database too.