div instead of p-tag for footer area
and showing an idea how a flyspray installation which shall be integrated with other website areas like a CMS, a wiki or something like that without the need to fiddle with the template files when upgrading flyspray.

naming $general_integration and $footer_integration are just ideas at the moment. The source of these may be some entries in the prefs database table in future. And you keep your CSS customization in a custom_*. css file..