die friendlier when not logged in
When someone is not logged in - but uses for instance a bookmark or a link in an email (e.g. example.com/proj123 or example.com/index.php?project=123 to a project that is visible only for logged-in users with permission - Flyspray just died with a blank page. 

By redirecting it shows at least the index page and the error message so the user can log in.