fallback to default theme .tpl files for incomplete themes
Enables writing themes without copying the full theme files from CleanFS-folder.

This enables to enhance Flyspray if a custom_*.css (colors, paddings, etc) or admin settings (footer areas etc. ) are not enough for customization, but keeps the redundancy of maintaining an extra theme low, because the main theme CleanFS is updated with each Flyspray release.