limit access to vendor/ directory
deny sniffing in the vendor/ directory.

Let's put it here (persistent) and not inside a vendor/.htaccess because 'composer' creates/manages vendor/ directory

Maybe the 3 files for the captcha are served in future by Flyspray, not by securimage folders. So maybe in future simple full deny of vendor/ directory access.