warn about bad utf8 old mysql version at install
If mysql(i) is choosen as database type and database not exists yet, Flyspray installer tries to create that with the right default character set utf8mb4 and collation utf8mb4_unicode_ci. But if thats not possible due hosting limitations, at least show a warning that only up to 3byte utf8 is supported my that mysql version, while 4byte utf8 is not and so full utf8 is only supported since mysql 5.5.3 with that 'utf8mb4'.

Applies only to mysql, postgresql not affected.