add massops checkbox in admin settings
enable/disable mass operations. Currently only for admins, later for project admins (old and target projects permissions must checked).
This can be seen as a temporary setting. Ideally it is calculated only on the global and project users permissions, which mass operations are possible and shown to the user in the frontend.