compare can_self_register() and can_register()
aligned the compared preference values in same order to make the differences between can_register() and can_self_register() more obvious. The naming of this 2 methods and the preference 'spam_proof' is probably not good, so added at least some documentation of what it really means currently. Also 'need_approval' (by an admin) setting is in my opinion independent of user email verification setting ('spam_proof' *sic*), so this is probably not the last change on this topic.