add tpl_tag()
tpl_tag() accepts also a simple #rgb or #rrggbb value set for the background-color instead of a named css class. Makes setting a simple background for a tag easy as no need to fiddle with a custom css file.

Also some phpdoc. Remove of folding markers ({{{}}}) as this can be done automatically by editors.