limit preload task description on long tasklists (default theme)
The default CleanFS theme loads also the task description of each task in a tasklist and only a mouse hover a task summary shows the description preview of a task. This feature has impact on required php workload and those descriptions are not used in most cases and also effects page size for each tasklist view. As compromise this change loads only the task description for maximal first 25 rows of a tasklist.