prevent flicker on tasklist
The previous caret bitmap image placed at background of a table cell could use the default cell CSS paddings to display. With the fontawesome caret icon for easier theme coloring this works different resulting in a bit jumpy tasklist rendering as the table is first rendered in (at least) Firefox 70 and then the font icon is shown which requires 9px where the cell content box is usually only 0 px wide for that column. By setting the content to 10px and reducing cell padding of caret column I try to avoid a second visual annoying tasklist render to the user.