fix FS#2591 not replace + at end of a search word
Flyspray has a hidden/undocumented feature since 2006 by using the + char in the search string for more narrow search conditions.
But if the + is at end of the search string, the replacement is unwanted/unneeded and a searchword like 'memtest86+' searched for 'memtest86 ' (with the space at end) so would not find 'memtest86+'. This patch takes care of it.