edit and view links for userlist in admin area
The links there pointed always to the edituser page. This was inconsistent with the other user links and annoyed me much. Now there is a separate edit user icon column and the links on the user_name and real_name point to the user page as the rest of Flyspray. It will surely not receive a design award but is a tiny bit improvement in the user management admin area.