Flyspray/Flyspray #880
Allow new tasks to be added in sequence by non-anonymous users.

In scripts/newtask.php:

Add a boolean value "addanothertask" to template data.  If the user is
logged in and the checkbox was previously checked, this state is
carried to the next new task via $_SESSION['addanothertask'].

If present, previously set session flag is cleared.

In themes/CleanFS/templates/newtask.tpl:

Add a checkbox "addanothertask", only visible to logged-in users.

In includes/modify.in.php, case "newtask.newtask":

If the "addanothertask" checkbox was not checked, redirect to task
details as before.

Otherwise, set $_SESSION['addanothertask'] = 1 and redirect
back to newtask.

On error, clear the session flag to prevent it from lingering.