Flyspray/Flyspray #887: Task List Table Header Sort Icons
Restored visibility of Attachment, Comments, and Votes icons which a
previous commit in FS_887_ui_improvements has broken.

Replace sort indicator images with Font Awesome Icons.

Wrapped colunn name text with span in order to selectively apply
margin-left to certain icons.

Primary sort icon is contained in the link as before.

Secondary sort icon (order2) added after the link as needed, colored
gray.  If Flyspray is going to carry forward the previous sort, the
user should have some indication that this is happening.

Sorted th elements now gain class 'orderby' or 'orderby2' as needed,
as well as class 'asc' or 'desc' to allow styling based on the current
state of the table.