Flyspray/Flyspray #887: Task Edit Add Comment Enhancements
In details.edit.tpl, replace the CSS checkbox sibling method used for
toggling the comment portion with an event handler in details.js.

detailsInit() gains and else block for this purpose.

The label is now a button, with a non-stacked comment icon.

When opening the comment portion, focus us automatically sent to the
textarea, which becomes enabled.

When closing the comment portion, if the trimmed textarea value is not
an empty string, the icon in the toggle button turns red.  Otherwise,
the icon color is restored.

Also when closing the comment portion, the textarea is disabled to
prevent the input being included in the POST body, avoiding creation
of unintended comments.