add audit ignores for some dependencies
composer prevents now installing dependencies for package versions with known security issues.

Ideally we would just use the current fixed versions of that packages, but this is not done yet. (extra migration work)

So temporarly added that ignore section to composer.json, because I think the vulnerable parts of that packages versions are not used or accessible by Flyspray.